+++ Permissions.js
@@ -1,24 +1,6 @@
 import GDrive from "./GDrive";
-import {
-   StaticUtils,
-   ArrayStringifier
-} from "simple-common-utils";
 
 const permissions = "/permissions";
-
-function _stringifyQueryParams(queryParams,
-   prefix = "?", separator = "&", quoteIfString)
-{
-   const array = [];
-   
-   Object.keys(queryParams).forEach(key => array.push(
-      `${key}=${StaticUtils.safeQuoteIfString(queryParams[key], quoteIfString)}`));
-   
-   return new ArrayStringifier(array)
-      .setPrefix(prefix)
-      .setSeparator(separator)
-      .process();
-}
 
 export default class Permissions {
    create(fileId, params, queryParams) {
@@ -26,7 +8,7 @@ export default class Permissions {
       
       let finalQueryParams = '';
       if (queryParams) {
-         finalQueryParams = _stringifyQueryParams(queryParams);
+         finalQueryParams = GDrive._stringifyQueryParams(queryParams);
       }
 
       return fetch(`${GDrive._urlFiles}/${fileId}${permissions}${finalQueryParams}`, {
